% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccardAndPValue.R
\name{jaccardAndPValue}
\alias{jaccardAndPValue}
\title{Ranks bed file similarities with significance}
\usage{
jaccardAndPValue(n, bed1, genome, folder_dir)
}
\arguments{
\item{n}{he number of background files generated (the greater the n, the more reliable the p-value, default n of 100)}

\item{bed1}{The file path string of a query bed file to be compared to the database files.}

\item{genome}{The file path of a genome file, which should be tab delimited and structured as follows: <chromName><TAB><chromSize>}

\item{folder_dir}{The directory of a folder containing separate database files to be used for comparison with the query file.}
}
\value{
A dataframe with 12 columns, ranked by jaccard index. Contains the name of the database file, the respective jaccard index, the pi score, the significance value (p-value) of the similarity, the proportion of intersection similarity with bed1, the proportion of intersection similarity with the database file, and a five-number summary of the background files generated from the query file.
}
\description{
The function compares a given query bed file with multiple bed files, and ranks the relative similarity between each file pairing, computed using jaccard indexes, where 0 has no similarities and 1 has an identical file. Will provide significance values (p-values) of each jaccard index, but will run slower.
}
