% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccardOnly.R
\name{jaccardOnly}
\alias{jaccardOnly}
\title{Ranks bed file similarities without significance}
\usage{
jaccardOnly(bed1, folder_dir)
}
\arguments{
\item{bed1}{The file path string of a query bed file to be compared to the database files.}

\item{folder_dir}{The file path string of a query bed file to be compared to the database files.}
}
\value{
A dataframe with four columns, ranked by jaccard index. Contains the name of the database file, the respective jaccard index, the proportion of intersection similarity with bed1, and proportion of intersection similarity with the database file.
}
\description{
The function compares a given query bed file with multiple bed files, and ranks the relative similarity between each file pairing, computed using jaccard indexes, where 0 has no similarities and 1 has an identical file. Will not provide significance values, but will run faster.
}
\examples{
jaccardOnly("/dir/bed1.txt","/dir/folder_dir")
}
