% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankBedSimilarity.R
\name{rankBedSimilarity}
\alias{rankBedSimilarity}
\title{Rank bed file similarities with or without significance}
\usage{
rankBedSimilarity(
  n = 100,
  bed1,
  genome,
  folder_dir,
  method = c("jaccard_pval", "jaccard_only")
)
}
\arguments{
\item{n}{The number of background files generated (the greater the n, the more reliable the p-value, default n of 100)}

\item{bed1}{The file path string of a query bed file to be compared to the database files}

\item{genome}{The file path of a genome file, which should be tab delimited and structured as follows: <chromName><TAB><chromSize>}

\item{folder_dir}{The directory of a folder containing separate database files to be used for comparison with the query file.}

\item{method}{The method that specifies the output format. "jaccard_pval" will output the jaccard indexes and the p-value significance. "jaccard_only" will only output jaccard indexes, but will run faster.}
}
\value{
A dataframe that shows the similarities of the query file to the database files ranked from greatest to least, tailored to user method specification.
}
\description{
The function compares a given query bed file with multiple bed files, and ranks the relative similarity between each file pairing, computed using jaccard indexes, where 0 has no similarities and 1 has an identical file. Can also provide a significance value (p-value) to the similarity index based on user selection.
}
